Added custome list routes

+++ app.js
@@ -16,7 +16,7 @@ mongoose.connect("mongodb://localhost:27017/todolistDB");
 const itemsSchema = new mongoose.Schema({
   name: {
     type: String,
-    required: true }
+    required: true },
 
 });
 
@@ -24,26 +24,33 @@ const Item = mongoose.model("item", itemsSchema);
 
 
 const item1 = new Item({
-  name: "Welcome to the To-Do List!"
+  name: "Welcome to the To-Do List!",
 
 });
 
 const item2 = new Item({
-  name: "Buy Food"
+  name: "Buy Food",
+
 
 });
 
 const item3 = new Item({
-  name: "Eat Food"
+  name: "Eat Food",
+
 });
 
 const defaultItems = [ item1, item2, item3 ];
 
+const listSchema = {
+  name: String,
+  items: [itemsSchema]
+}
 
+const List = mongoose.model("List", listSchema);
 
 app.get("/", function(req, res) {
 
-    Item.find().then(function(items){
+    Item.find({listName:'default'}).then(function(items){
       if (items.length === 0){ 
         Item.insertMany(defaultItems).then(function(){
           console.log("Successfully saved default items to DB");
@@ -52,47 +59,74 @@ app.get("/", function(req, res) {
         })
         res.redirect("/"); }
       else {
-        res.render("list", {listTitle: "Today", newListItems: items});
+        res.render("list", {listTitle: "today", newListItems: items});
       }
 
   }).catch(function(error){
   console.log(error);
   });
-
-
-
 });
 
 app.post("/", function(req, res){
 
   const itemName = req.body.newItem;
+  const listName = req.body.list;
 
   const item = new Item({
-    name: itemName
-  });
-  item.save();
-  res.redirect("/");
+    name: itemName,
+    });
+
+    if(listName === "Today"){
+      item.save();
+      res.redirect("/");
+    }
+    else {
+      List.findOne({ name: listName}).then(function(foundList){
+        foundList.items.push(item);
+        foundList.save();
+        res.redirect("/" + listName);
+      });
+
+    }
+
 
 });
 
 app.post("/delete", function(req, res){
   const checkedId =req.body.checkbox;
+  const listName = req.body.listName;
+
+  if (listName === "Today"){
   Item.findByIdAndDelete(checkedId).then(function(){
     console.log("Successfully deleted item")})
     res.redirect("/");
- 
-})
+  }
 
+  else {
+    List.findOneAndUpdate({name: listName}, {$pull: {items: {_id: checkedId}}}).then(function(){
+      res.redirect("/" + listName);
+    })};
 
-app.get("/:listTitle", (req,res) => {
-  const AnotherItem = mongoose.model("AnotherItem", itemsSchema);
-  const anotherItemName = new AnotherItem({
-    name: req.params.listTitle
-  })
-  res.redirect("/" + req.params.listTitle);
+  });
+app.get("/:customListName", (req,res) => {
+  const customListName = req.params.customListName;
+
+  List.findOne({name: customListName}).then(function(foundList){
+    if(!foundList){
+      const list = new List ({
+        name: customListName,
+        items: defaultItems
+      });
+      
+      list.save();
+      res.redirect("/" + customListName);
+        }   
+      else {
+        res.render("list", {listTitle: customListName, newListItems: foundList.items});
+
+      }
 });
-  res.render("list", {listTitle: req.params.listTitle, newListItems: defaultItems});
-
+});
 
 
 app.get("/about", function(req, res){
